requirements.txt
pandas
streamlit
opencv-python
numpy
insightface
onnxruntime
faiss-cpu
ultralytics
torch
torchvision
sqlalchemy
streamlit-autorefresh
filelock
webrtcvad
openai-whisper



